resource split
