+++ deeplearning4j-ui/src/main/java/org/deeplearning4j/ui/nearestneighbors/word2vec/NearestNeighborsResource.java
@@ -63,12 +63,6 @@ public class NearestNeighborsResource extends FileResource {
 
     }
 
-    @GET
-    public View get() {
-        return new NearestNeighborsView();
-    }
-
-
     @POST
     @Path("/vocab")
     @Produces(MediaType.APPLICATION_JSON)
